chore(tangle-cloud): rename services hooks
