+++ apps/tangle-cloud/src/components/InstanceHeader.tsx
@@ -20,7 +20,8 @@ const InstanceHeader: FC<Props> = ({ title, githubPath, creator }) => {
     >
       <div className="flex flex-col gap-2">
         <Typography variant="h5" className="flex items-center gap-2">
-          {title} <GithubFill size="lg" target="_blank" href={githubPath || '#'} />
+          {title}{' '}
+          <GithubFill size="lg" target="_blank" href={githubPath || '#'} />
         </Typography>
 
         <Typography variant="body1" className="!text-mono-100">
